requirements.txt
requests>=2.18.2
PyYAML>=3.13
